fix(core): generate prompt and nx list should support encapsulated repos
