+++ packages/nx/src/utils/plugins/index.ts
@@ -4,7 +4,7 @@ export {
 } from './community-plugins';
 export { fetchCorePlugins, listCorePlugins } from './core-plugins';
 export {
-  getInstalledPluginsFromPackageJson,
+  getInstalledPluginsAndCapabilities,
   listInstalledPlugins,
 } from './installed-plugins';
 export {
