+++ packages/nx/src/utils/plugins/installed-plugins.ts
@@ -4,20 +4,77 @@ import type { CommunityPlugin, CorePlugin, PluginCapabilities } from './models';
 import { getPluginCapabilities } from './plugin-capabilities';
 import { hasElements } from './shared';
 import { readJsonFile } from '../fileutils';
-import { readModulePackageJson } from '../package-json';
+import { PackageJson, readModulePackageJson } from '../package-json';
+import { workspaceRoot } from '../workspace-root';
+import { join } from 'path';
+import { NxJsonConfiguration } from '../../config/nx-json';
 
-export function getInstalledPluginsFromPackageJson(
+export function findInstalledPlugins(): PackageJson[] {
+  const packageJsonDeps = getDependenciesFromPackageJson();
+  const nxJsonDeps = getDependenciesFromNxJson();
+  const deps = packageJsonDeps.concat(nxJsonDeps);
+  return deps.reduce(
+    (arr: any[], nextDep: string): { project: string; version: string }[] => {
+      try {
+        const depPackageJson: Partial<PackageJson> =
+          readModulePackageJson(nextDep, [
+            workspaceRoot,
+            join(workspaceRoot, '.nx', ' installation'),
+          ]).packageJson || {};
+        if (
+          [
+            'ng-update',
+            'nx-migrations',
+            'schematics',
+            'generators',
+            'builders',
+            'executors',
+          ].some((field) => field in depPackageJson)
+        ) {
+          arr.push({ package: nextDep, ...depPackageJson });
+          return arr;
+        } else {
+          return arr;
+        }
+      } catch {
+        return arr;
+      }
+    },
+    []
+  );
+}
+
+function getDependenciesFromPackageJson(
+  packageJsonPath = 'package.json'
+): string[] {
+  try {
+    const { dependencies, devDependencies } = readJsonFile(
+      join(workspaceRoot, packageJsonPath)
+    );
+    return Object.keys({ ...dependencies, ...devDependencies });
+  } catch {}
+  return [];
+}
+
+function getDependenciesFromNxJson(): string[] {
+  const { installation } = readJsonFile<NxJsonConfiguration>(
+    join(workspaceRoot, 'nx.json')
+  );
+  if (!installation) {
+    return [];
+  }
+  return ['nx', ...Object.keys(installation.plugins || {})];
+}
+
+export function getInstalledPluginsAndCapabilities(
   workspaceRoot: string,
   corePlugins: CorePlugin[],
   communityPlugins: CommunityPlugin[] = []
 ): Map<string, PluginCapabilities> {
-  const packageJson = readJsonFile(`${workspaceRoot}/package.json`);
-
   const plugins = new Set([
     ...corePlugins.map((p) => p.name),
     ...communityPlugins.map((p) => p.name),
-    ...Object.keys(packageJson.dependencies || {}),
-    ...Object.keys(packageJson.devDependencies || {}),
+    ...findInstalledPlugins().map((p) => p.name),
   ]);
 
   return new Map(
